requirements.txt
streamlit==1.24.0
google-generativeai==0.1.0
python-dotenv==0.19.0